requirements.txt
anthropic>=0.40.0
tavily-python>=0.5.0
agentmail>=0.1.0
python-dotenv>=1.0.0
